+++ src/main/java/com/github/estebanzuniga/finalreality/model/character/player/PlayerCharacter.java
@@ -1,19 +1,23 @@
 package com.github.estebanzuniga.finalreality.model.character.player;
 
 import com.github.estebanzuniga.finalreality.model.character.AbstractCharacter;
+import com.github.estebanzuniga.finalreality.model.character.CharacterClass;
 import com.github.estebanzuniga.finalreality.model.character.ICharacter;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 
+import com.github.estebanzuniga.finalreality.model.weapon.Weapon;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * A class that holds all the information of a single character of the game.
  *
  * @author Ignacio Slater Muñoz.
- * @author <Your name>
+ * @author Esteban Zúñiga Salamanca.
  */
 public class PlayerCharacter extends AbstractCharacter {
+
+  private Weapon equippedWeapon = null;
 
   /**
    * Creates a new character.
@@ -26,9 +30,17 @@ public class PlayerCharacter extends AbstractCharacter {
    *     the class of this character
    */
   public PlayerCharacter(@NotNull String name,
-      @NotNull BlockingQueue<ICharacter> turnsQueue,
-      final CharacterClass characterClass) {
+                         @NotNull BlockingQueue<ICharacter> turnsQueue,
+                         final CharacterClass characterClass) {
     super(turnsQueue, name, characterClass);
+  }
+
+  public void equip(Weapon weapon) {
+    this.equippedWeapon = weapon;
+  }
+
+  public Weapon getEquippedWeapon() {
+    return equippedWeapon;
   }
 
   @Override
@@ -46,6 +58,6 @@ public class PlayerCharacter extends AbstractCharacter {
     }
     final PlayerCharacter that = (PlayerCharacter) o;
     return getCharacterClass() == that.getCharacterClass()
-        && getName().equals(that.getName());
+            && getName().equals(that.getName());
   }
 }