+++ src/main/java/com/github/estebanzuniga/finalreality/model/weapon/Weapon.java
@@ -6,7 +6,7 @@ import java.util.Objects;
  * A class that holds all the information of a weapon.
  *
  * @author Ignacio Slater Muñoz.
- * @author <Your name>
+ * @author Esteban Zúñiga Salamanca.
  */
 public class Weapon {