+++ src/main/java/com/github/estebanzuniga/finalreality/model/weapon/WeaponType.java
@@ -4,7 +4,7 @@ package com.github.estebanzuniga.finalreality.model.weapon;
  * Enumeration of all the weapon types.
  *
  * @author Ignacio Slater Muñoz.
- * @author <Your name>
+ * @author Esteban Zúñiga Salamanca.
  */
 public enum WeaponType {
   SWORD, AXE, KNIFE, STAFF, BOW