+++ src/test/java/com/github/cc3002/finalreality/model/character/EnemyTest.java
@@ -1,7 +1,7 @@
 package com.github.cc3002.finalreality.model.character;
 
 import com.github.estebanzuniga.finalreality.model.character.Enemy;
-import com.github.estebanzuniga.finalreality.model.character.player.CharacterClass;
+import com.github.estebanzuniga.finalreality.model.character.CharacterClass;
 import com.github.estebanzuniga.finalreality.model.character.player.PlayerCharacter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;