+++ src/test/java/com/github/cc3002/finalreality/model/character/PlayerCharacterTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 import com.github.estebanzuniga.finalreality.model.character.Enemy;
-import com.github.estebanzuniga.finalreality.model.character.player.CharacterClass;
+import com.github.estebanzuniga.finalreality.model.character.CharacterClass;
 import com.github.estebanzuniga.finalreality.model.character.player.PlayerCharacter;
 import java.util.EnumMap;
 import java.util.Map;