+++ src/test/java/com/github/cc3002/finalreality/model/weapon/WeaponTest.java
@@ -1,11 +1,18 @@
 package com.github.cc3002.finalreality.model.weapon;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
+import com.github.estebanzuniga.finalreality.model.character.ICharacter;
+import com.github.estebanzuniga.finalreality.model.character.CharacterClass;
+import com.github.estebanzuniga.finalreality.model.character.player.PlayerCharacter;
 import com.github.estebanzuniga.finalreality.model.weapon.Weapon;
 import com.github.estebanzuniga.finalreality.model.weapon.WeaponType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.util.concurrent.BlockingQueue;
+import java.util.concurrent.LinkedBlockingQueue;
 
 class WeaponTest {
 
@@ -22,6 +29,8 @@ class WeaponTest {
   private Weapon testSword;
   private Weapon testBow;
   private Weapon testKnife;
+
+  protected BlockingQueue<ICharacter> turns = new LinkedBlockingQueue<>();
 
   @BeforeEach
   void setUp() {
@@ -40,6 +49,12 @@ class WeaponTest {
     var expectedBow = new Weapon(BOW_NAME, DAMAGE, SPEED, WeaponType.BOW);
     var expectedKnife = new Weapon(KNIFE_NAME, DAMAGE, SPEED, WeaponType.KNIFE);
 
+    var expectedCharacter = new PlayerCharacter("Test", turns, CharacterClass.KNIGHT);
+    var difNameExpectedAxe = new Weapon("Not Test Axe", DAMAGE, SPEED, WeaponType.AXE);
+    var difDamageExpectedAxe = new Weapon(AXE_NAME, DAMAGE+1, SPEED, WeaponType.AXE);
+    var difSpeedExpectedAxe = new Weapon(AXE_NAME, DAMAGE, SPEED+1, WeaponType.AXE);
+    var difTypeExpectedWeapon = new Weapon(AXE_NAME, DAMAGE, SPEED, WeaponType.SWORD);
+
     assertEquals(expectedAxe, testAxe);
     assertEquals(expectedAxe.hashCode(), testAxe.hashCode());
     assertEquals(expectedStaff, testStaff);
@@ -50,5 +65,13 @@ class WeaponTest {
     assertEquals(expectedBow.hashCode(), testBow.hashCode());
     assertEquals(expectedKnife, testKnife);
     assertEquals(expectedKnife.hashCode(), testKnife.hashCode());
+
+    assertNotEquals(expectedAxe, expectedCharacter);
+    assertNotEquals(expectedAxe.hashCode(), expectedCharacter.hashCode());
+
+    assertNotEquals(expectedAxe, difNameExpectedAxe);
+    assertNotEquals(expectedAxe, difDamageExpectedAxe);
+    assertNotEquals(expectedAxe, difSpeedExpectedAxe);
+    assertNotEquals(expectedAxe, difTypeExpectedWeapon);
   }
 }